pyproject.toml
[project]
name = "proper_pixel_art"
version = "1.3.0"
description = "Converts pixel-art-style images such as those from generative models or low-quality sprite web uploads to true resolution usable assets."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
]
packages = [
    { include = "proper_pixel_art" }
]

[project.scripts]
cli = "proper_pixel_art.cli:main"
ppa = "proper_pixel_art.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.4.2",
    "ruff>=0.14.4",
]
